chatgpt_crypto_ai/requirements.txt
flask==2.3.3
gunicorn==20.1.0
ccxt
openai==1.3.0
python-dotenv==1.0.0
Flask-Cors==4.0.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.4
python-jose==3.3.0
redis==4.6.0
PyJWT==2.10.1
tenacity
